requirements.txt
asgiref==3.8.1
babel==2.16.0
beautifulsoup4==4.12.3
boto3==1.35.66
botocore==1.35.66
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
Django==4.2.16
django-crispy-forms==2.3
django-extra-views==0.14.0
django-haystack==3.3.0
django-modeltranslation==0.19.12
django-oscar==3.2.5
django-phonenumber-field==6.4.0
django-storages==1.14.4
django-tables2==2.3.4
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
factory-boy==3.2.1
Faker==30.8.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
numpy==2.0.2
packaging==24.1
pandas==2.2.3
phonenumbers==8.13.48
pillow==11.0.0
psycopg2-binary==2.9.10
purl==1.6
pycountry==24.6.1
pysolr==3.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.4
setuptools==75.3.0
six==1.16.0
sorl-thumbnail==12.9.0
soupsieve==2.6
sqlparse==0.5.1
transliterate==1.10.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.10
whitenoise==6.8.2
